+++ apps/consent-ui/src/components/common/Modal/index.tsx
@@ -27,6 +27,7 @@ import Button from '../Button';
 
 import styles from './Modal.module.scss';
 
+// props needed when using Modal in a component ex. RegisterMinorModal
 export type ModalComponentProps = {
 	closeModal: () => void;
 	currentLang: ValidLanguage;
